custom houses fixes

+++ server/src/network/uotxpackets.h
@@ -1290,7 +1290,7 @@ public:
 	void setCompression( UINT8 data ) { (*this)[3] = data; }
 	void setSerial( UINT32 data ) { setInt( 5, data ); }
 	void setRevision( UINT32 data ) { setInt( 9, data ); }
-	void addTile( UINT16 id, UINT8 x, UINT8 y, UINT8 z );
+	void addTile( UINT16 id, short x, short y, short z );
 	void addTile( UINT16 id, Coord_cl coords ) { addTile( id, coords.x, coords.y, coords.z ); }
 };
